pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Zarnigor <zarnigorn1255@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
uvicorn = "^0.31.0"
jinja2 = "^3.1.4"
sqladmin = "^0.19.0"
psycopg2-binary = "^2.9.9"
sqlmodel = "^0.0.22"
alembic = "^1.13.3"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
databases = "^0.9.0"
python-slugify = "^8.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
